refactor: calculate win
